BUG 8140 : Elect nodes for subnetroute on control path reboot
When a physical switch on the cloud management network in the cloud is
rebooted, it results in disconnection of datapath-nodes from the ODL
controller. Under such conditions, the subnetRoute is removed from all
the DPNs.
However, once the physical switch comes back up after reboot, the
datapath-nodes are all getting reconnected to the ODL controller.  During
this time in a scaled environment, the subnetRoute flow rules aren't
getting reapplied to the datapath-nodes.  This is because the election
process fails in figuring out a designated datapath-node for hosting
subnetroute, thereby the subnetRoute is not getting reconfigured on all
the datapath-nodes.

+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/BgpManager.java
@@ -134,7 +134,7 @@ public class BgpManager implements AutoCloseable, IBgpManager {
 
     @Override
     public void advertisePrefix(String rd, String macAddress, String prefix, List<String> nextHopList,
-                                VrfEntry.EncapType encapType, int vpnLabel, long l3vni, long l2vni,
+                                VrfEntry.EncapType encapType, long vpnLabel, long l3vni, long l2vni,
                                 String gatewayMac) throws Exception {
         bcm.addPrefix(rd, macAddress, prefix, nextHopList,
                 encapType, vpnLabel, l3vni, l2vni, gatewayMac, 1 /* TODO FIX afi */);
@@ -142,7 +142,7 @@ public class BgpManager implements AutoCloseable, IBgpManager {
 
     @Override
     public void advertisePrefix(String rd, String macAddress, String prefix, String nextHop,
-                                VrfEntry.EncapType encapType, int vpnLabel, long l3vni, long l2vni,
+                                VrfEntry.EncapType encapType, long vpnLabel, long l3vni, long l2vni,
                                 String gatewayMac) throws Exception {
         LOG.info("ADVERTISE: Adding Prefix rd {} prefix {} nexthop {} label {}", rd, prefix, nextHop, vpnLabel);
         bcm.addPrefix(rd, macAddress, prefix, Collections.singletonList(nextHop), encapType,
